版本失效恢复,几乎所有文件同步状态全丢,被迫再次同步1k9个文件(
变化:更好的combo game(acc score比值甚至高达0.00001%呢(不是))
改进尾迹显示(更短的长尾迹显示时长,类似ATRI的某项功能)
删除大量娱乐mod并把relax扔到ez模组列表(这个是没事干(((((

+++ osu.Game.Tests/Visual/UserInterface/TestSceneSectionsContainer.cs
@@ -138,7 +138,7 @@ namespace osu.Game.Tests.Visual.UserInterface
                 AddUntilStep("section top is visible", () =>
                 {
                     var scrollContainer = container.ChildrenOfType<UserTrackingScrollContainer>().Single();
-                    double sectionPosition = scrollContainer.GetChildPosInContent(container.Children[scrollIndex]);
+                    float sectionPosition = scrollContainer.GetChildPosInContent(container.Children[scrollIndex]);
                     return scrollContainer.Current < sectionPosition;
                 });
             }
